re-organized build scripts
